refactor(team): explicit wiring layer + headless recovery split (C18)
Three semantic regroupings, applying patterns proven in earlier PRs:

1. **Explicit wiring layer**: launcher_wiring.go (191 lines, new)
   collects the five lazy sub-type accessors (targeter, notifyCtx,
   scheduler, paneDispatch, panes) plus the two scheduler thin
   wrappers (updateSchedulerJob, watchdogSchedulerLoop). Every
   accessor follows the same shape — nil-safe early return,
   one-shot construction stored on the launcher, callbacks captured
   by closure so the sub-type sees current Launcher state without
   holding a Launcher pointer. This file is the orchestrator's
   wiring layer, made explicit and isolated from construction
   (NewLauncher) and lifecycle (Launch/Kill).

2. **Helpers next to their state**: recordPaneSpawnFailure moved
   to launcher_membership.go where the failedPaneSlugs map's other
   touch sites live (officeMembersSnapshot, agentConfigFromMember,
   etc). HasLiveTmuxSession moved to pane_lifecycle.go where the
   paneLifecycle type lives — it's the package-level free-function
   entry point for the same code path.

3. **Headless recovery split**: headless_codex_recovery.go (~400
   lines, new) extracts the post-turn recovery layer:
   - durability checks (taskHasDurableCompletionState,
     headlessTurnCompletedDurably,
     taskHasExternalWorkflowEvidenceSince)
   - retry-prompt builders (headlessTimedOutRetryPrompt,
     headlessFailedRetryPrompt, shouldRetryHeadlessTurn,
     shouldRetryTimedOutHeadlessTurn)
   - recovery dispatchers (recoverTimedOutHeadlessTurn,
     recoverFailedHeadlessTurn, timedOutTaskForTurn,
     timedOutTurnAlreadyRecovered)
   - agent-posting heuristics (agentPostedSubstantiveMessageSince,
     agentPostedSubstantiveMessageToChannelSince,
     postHeadlessFinalMessageIfSilent, headlessReplyToID)
   - durability-failure helper (isDurabilityFailure)

   Mirrors the C10 split pattern: headless_codex.go is now entry
   points + types only; queue/runner/recovery each own a sibling
   file.

launcher.go: 579 -> 382 lines (-197).
headless_codex.go: 683 -> 297 lines (-386).

Cumulative since main: 4998 -> 382 = -4616 lines (-92.4%).

What's left in launcher.go (382 lines): genuinely the orchestrator
core — package init (port-suffixed session names), the Launcher
struct definition, NewLauncher construction, claudeCommand (held
back from C13 by the no-secrets hook false positive), and the
const block for agent notify cooldowns.

Stacked on refactor/launcher-c17-final-grooming (C17).
